Switch to `aws_s3_bucket_lifecycle_configuration` (#463)
* Switch to `aws_s3_bucket_lifecycle_configuration`

We are seeing Terraform attempting to remove the following on every run:
      - server_side_encryption_configuration {
          - rule {
              - bucket_key_enabled = false -> null

              - apply_server_side_encryption_by_default {
                  - sse_algorithm = "AES256" -> null
                }
            }
        }

Upgrading the AWS Provider to v4 should fix this.
Before we can upgrade to v4 we need to switch to using
  aws_s3_bucket_lifecycle_configuration

* Move AWS provider version into required_providers block

Previous declaration was deprecated:
```
Terraform 0.13 and earlier allowed provider version constraints inside the
provider configuration block, but that is now deprecated and will be removed
in a future version of Terraform. To silence this warning, move the provider
version constraint into the required_providers block.
```

* Bump version of Terraform for GP2GP module

The list function has been deprecated since 0.12 [1], so replace with
tolist.

[1]: https://developer.hashicorp.com/terraform/language/functions/list

* Upgrade GP2GP AWS Terraform provider to version 4

Fixes issue where Terraform complains about S3 lifecycle rules
needing updating when no terraform has changed.